+++ resources/views/components/combo-line-column.blade.php
@@ -1,0 +1,5 @@
+@props(['series' => [], 'categories' => [], 'legend' => true])
+
+<x-chart::base component="combo-line-column" :option="\WireCharts\Support\Option::comboLineColumn([
+    'series' => $series, 'categories' => $categories, 'legend' => $legend,
+])" {{ $attributes }} />